Lots of fixes & Create Group Now Working
Fixes include table headers not being in bold (hadn't downloaded roboto 700)
Fixed a few timezone issues.
Started writing up the function that will format duration from unix timestamp to human comprehendable time.
Updated the title of most pages since I forgot to do that when copying the HTML to form a new page with identical styling.
Created code to allow the user to join a group from the table.
Created code to allow the user to create a group.
Consolidated code on all loged in pages to detect whether or not the user is logged in or not and redirect them to a log in page if the latter is applicable.