Lots of Small Changes
Created time formatting function.
Got time formatting on the group creation for cycle duration to snap to the nearest hour.
Edited code so that, on the 'Dashboard' page it pulls data from the 'Groups' table to fill in name and got time since last death working.
Made Login more efficient and secure.