Dashboard Information Upgrade + Group Page Creation + Small Fixes
Vastly increased the capabilities of the information in the table displayed on the Dashboard page, including the length of time since the last death event, the required pay out and potential winnings from an immenent death. And a button to go to the group.

The Group Page relies on $_GET value in the URL to retrieve the ID of the group that is being accessed. A check is done with the Group ID and the User ID (from the session) to ensuer that the user is a member of the group in question, otherwise redirects back to Dashboard.

small chages include the addition of a semicolon on the Greate Groups page, and realigning of tag tabbing in the Join Groups page