Lots of Changes
Created a testing page to test aspects involved with connecting to the database with python.

Moved the post validation section of the join groups page above all of the other PHP such that an error doesn't occur when it would appear that the user wasnt a part of the group and then they could join twice... Also made a section of validation such that it double checks to see if the user is already a member of the group incase the page reloads and re-posts the values without having to press a button.

Started work on the tables that will be displayed on the group page.

Added a bit of formatting to the table on the dashboard page such that it will give a total of the number of groups that they are a part of.

Added the maximim selection of celebrities to the form when creating a group and added all of the respective logic and altered the SQL query to add it.