Fixed Spradic Error in which log in page would not log you in until the second attempt and Added far more content to the Group Page page.
Moved the php to test for the log in to the first thing that loads...

Made it such that the group page outputs a lot of stuff and gives the user the option to input a new celebrity and submit it, but the submission is yet to be implemented. Furthermore, added the code to display all the other selections from everyone else in the group.