Bringing Python Together
The Python is still not liking me and habing errors, but the new file 'DataRetrievalAndProcessing.py' is what i am using to drag and drop the code from my other python files to create the final file.

Got the Group Page to add the celebrity if required and to add the celebrity, group and user IDs to the Selection table in order to make their selection.